reorg to make more general beyond just clinical manifestations
