+++ src/clinical_status_manager.rs
@@ -4,15 +4,15 @@ use std::{
 };
 
 use ixa::{
-    define_data_plugin, define_rng, Context, ContextPeopleExt, ContextRandomExt, PersonProperty,
-    PersonPropertyChangeEvent,
+    define_data_plugin, define_rng, warn, Context, ContextPeopleExt, ContextRandomExt, IxaError,
+    PersonProperty, PersonPropertyChangeEvent,
 };
 
-define_rng!(ClinicalRng);
+define_rng!(ProgressionRng);
 
 /// Defines a semi-Markovian method for getting the next person property value based on the last.
 pub trait PropertyProgression {
-    /// The value of the person property that is being tracked by the clinical progression.
+    /// The value of the person property that is being tracked by the progression.
     type Value;
     /// Returns the next value and the time to the next value given the current value and `Context`.
     fn next(&self, context: &Context, last: &Self::Value) -> Option<(Self::Value, f64)>;
@@ -58,7 +58,7 @@ impl ContextPropertyProgressionExt for Context {
                 let progressions = container.progressions.get(&TypeId::of::<T>()).unwrap();
                 // Todo(kzs9): Make this not random but rather we pick the same index as the rate
                 // function id/some way of correlation between natural history
-                let id = context.sample_range(ClinicalRng, 0..progressions.len());
+                let id = context.sample_range(ProgressionRng, 0..progressions.len());
                 let tcr = progressions[id]
                     .downcast_ref::<Box<dyn PropertyProgression<Value = T::Value>>>()
                     .unwrap()
@@ -74,19 +74,75 @@ impl ContextPropertyProgressionExt for Context {
     }
 }
 
+/// Holds a sequence of unique states and times between subsequent states for changing a person's
+/// person property values accordingly. Since all person properties that have Markovian transitions
+/// may want to have their progressions defined empirically, we make this a general struct that can
+/// hold any type `T` that implements `PartialEq` and `Copy`.
+pub struct EmpiricalProgression<T: PartialEq + Copy> {
+    states: Vec<T>,
+    time_to_next: Vec<f64>,
+}
+
+impl<T: PartialEq + Copy> EmpiricalProgression<T> {
+    /// Makes a new `EmpiricalProgression<T>` struct that holds a sequence of states of value `T`.
+    /// Assumes values in `states` are unique.
+    /// # Errors
+    /// - If `states` is not one element longer than `time_to_next`.
+    pub fn new(
+        states: Vec<T>,
+        time_to_next: Vec<f64>,
+    ) -> Result<EmpiricalProgression<T>, IxaError> {
+        if states.len() != time_to_next.len() + 1 {
+            return Err(IxaError::IxaError(
+                "Size mismatch: states must be one element longer than time_to_next. Instead, "
+                    .to_string()
+                    + &format!(
+                        "states has length {} and time_to_next has length {}.",
+                        states.len(),
+                        time_to_next.len()
+                    ),
+            ));
+        }
+        warn!(
+            "Adding an EmpiricalProgression. At this time, we do not check whether values in
+        states are unique."
+        );
+        Ok(EmpiricalProgression {
+            states,
+            time_to_next,
+        })
+    }
+}
+
+impl<T: PartialEq + Copy> PropertyProgression for EmpiricalProgression<T> {
+    type Value = T;
+    fn next(&self, _context: &Context, last: &Self::Value) -> Option<(Self::Value, f64)> {
+        let mut iter = self.states.iter().enumerate();
+        while let Some((_, status)) = iter.next() {
+            if status == last {
+                return iter
+                    .next()
+                    .map(|(i, next)| (*next, self.time_to_next[i - 1]));
+            }
+        }
+        None
+    }
+}
 #[cfg(test)]
 mod test {
 
     use std::any::TypeId;
 
-    use ixa::{Context, ContextPeopleExt, ContextRandomExt, ExecutionPhase};
+    use ixa::{Context, ContextPeopleExt, ContextRandomExt, ExecutionPhase, IxaError};
 
     use crate::{
         population_loader::Age,
-        symptom_progression::{DiseaseSeverity, DiseaseSeverityValue, EmpiricalProgression},
+        symptom_progression::{DiseaseSeverity, DiseaseSeverityValue},
     };
 
-    use super::{ContextPropertyProgressionExt, Progressions, PropertyProgression};
+    use super::{
+        ContextPropertyProgressionExt, EmpiricalProgression, Progressions, PropertyProgression,
+    };
 
     struct AgeProgression {
         time_to_next_age: f64,
@@ -100,7 +156,7 @@ mod test {
     }
 
     #[test]
-    fn test_register_clinical_progression_automates_moves() {
+    fn test_register_property_progression_automates_moves() {
         let mut context = Context::new();
         context.init_random(0);
         let symptom_progression = EmpiricalProgression::new(
@@ -218,5 +274,28 @@ mod test {
         assert!(tcr
             .next(&context, &Some(DiseaseSeverityValue::Moderate))
             .is_none());
+    }
+
+    #[test]
+    fn test_empirical_progression_errors_len() {
+        let progression = EmpiricalProgression::new(
+            vec![
+                DiseaseSeverityValue::Mild,
+                DiseaseSeverityValue::Moderate,
+                DiseaseSeverityValue::Mild,
+            ],
+            vec![1.0, 2.0, 3.0],
+        );
+        let e = progression.err();
+        match e {
+            Some(IxaError::IxaError(msg)) => {
+                assert_eq!(msg, "Size mismatch: states must be one element longer than time_to_next. Instead, states has length 3 and time_to_next has length 3.".to_string());
+            }
+            Some(ue) => panic!(
+                "Expected an error that states and time_to_next have incompatible sizes. Instead got {:?}",
+                ue.to_string()
+            ),
+            None => panic!("Expected an error. Instead, validation passed with no errors."),
+        }
     }
 }
